requirements.txt
anthropic>=0.40.0
pydantic>=2.6.0
rich>=13.7.0
click>=8.1.7
python-dotenv>=1.0.0
